+++ services/util/python/util-simulation-service/src/util_simulation_service/model/agent_mode.py
@@ -6,3 +6,4 @@ class AgentMode(Enum):
 
     MICROWEIGHT = "microweight"
     KIND = "KIND"
+    TEST = "test"