+++ services/util/python/util-simulation-service/src/util_simulation_service/model/bootstrap_input.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+from util_simulation_service.model.domain_ontology_bootstrap_input import DomainOntologyBootstrapInput
+
+
+class BootstrapInput(BaseModel):
+    domain_ontologies: list[DomainOntologyBootstrapInput] = []